# internal/httpapi/session_handlers.go
package httpapi

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/aperture/aperture/internal/auth"
	"github.com/aperture/aperture/internal/browser"
	"github.com/aperture/aperture/internal/config"
	"github.com/aperture/aperture/internal/session"
	"github.com/gin-gonic/gin"
)

func toSessionResponse(view *session.SessionView) sessionResponse {
	browserArgs := make([]string, 0)
	_ = json.Unmarshal([]byte(view.Session.BrowserArgsJSON), &browserArgs)
	resp := sessionResponse{
		ID:               view.Session.ID,
		TenantID:         view.Session.TenantID,
		BaseSnapshotName: view.BaseSnapshotName,
		Label:            view.Session.Label,
		Status:           view.Session.Status,
		Browser: sessionBrowserConfig{
			Channel: view.Session.BrowserChannel,
			Mode:    view.Session.BrowserMode,
			Args:    browserArgs,
		},
		Capabilities: view.Capabilities,
		Connection: sessionConnection{
			CDPURL:       view.CDPURL,
			SessionToken: view.SessionToken,
		},
		CreatedAt:       view.Session.CreatedAt,
		StartedAt:       view.Session.StartedAt,
		StoppedAt:       view.Session.StoppedAt,
		DeletedAt:       view.Session.DeletedAt,
		ExpiresAt:       view.Session.ExpiresAt,
		LastConnectedAt: view.Session.LastConnectedAt,
		SuspendedAt:     view.Session.SuspendedAt,
		Tags:            view.Tags,
	}
	if resp.Connection.CDPURL != "" {
		for _, transport := range view.Capabilities.LiveView.Transports {
			if transport == browser.LiveViewTransportWebRTC {
				resp.Connection.WebRTC = &sessionWebRTCConnection{ICEServers: toICEServerResponses(view.WebRTCICEServers)}
				break
			}
		}
	}
	if view.CollaborationCapabilities.Editor != "" && view.CollaborationCapabilities.Viewer != "" {
		resp.Collaboration = &sessionCollaborationCapabilities{
			EditorToken: view.CollaborationCapabilities.Editor,
			ViewerToken: view.CollaborationCapabilities.Viewer,
		}
	}
	return resp
}

func (s *Server) rotateCollaborationCapability(c *gin.Context, role session.CollaborationRole) {
	if s.Sessions == nil {
		WriteError(c, errSessionServiceUnavailable)
		return
	}
	view, err := s.Sessions.RotateCollaborationCapability(c.Request.Context(), tenantIDFromContext(c), c.Param("sessionId"), role)
	if err != nil {
		WriteError(c, err)
		return
	}
	c.JSON(http.StatusOK, sessionMutationResponse{Session: toSessionResponse(view)})
}

func toICEServerResponses(servers []config.WebRTCICEServer) []iceServerResponse {
	if len(servers) == 0 {
		return nil
	}
	responses := make([]iceServerResponse, 0, len(servers))
	for _, server := range servers {
		responses = append(responses, iceServerResponse{
			URLs:       append([]string(nil), server.URLs...),
			Username:   server.Username,
			Credential: server.Credential,
		})
	}
	return responses
}

func (s *Server) createSession(c *gin.Context) {
	if s.Sessions == nil {
		WriteError(c, errSessionServiceUnavailable)
		return
	}

	var req createSessionRequest
	if err := bindJSON(c, &req); err != nil {
		WriteError(c, err)
		return
	}

	if req.BaseSnapshotName != nil && strings.TrimSpace(*req.BaseSnapshotName) != "" {
		principal := c.MustGet("principal").(auth.Principal)
		if !auth.HasScope(principal.Scopes, auth.ScopeSnapshotsRead) {
			WriteError(c, auth.ErrScopeDenied)
			return
		}
		if err := s.Auth.AuthorizeSnapshotName(c.Request.Context(), principal, tenantIDFromContext(c), *req.BaseSnapshotName); err != nil {
			WriteError(c, err)
			return
		}
	}

	view, err := s.Sessions.Create(c.Request.Context(), session.CreateInput{
		TenantID:         tenantIDFromContext(c),
		BaseSnapshotName: req.BaseSnapshotName,
		Label:            req.Label,
		BrowserChannel:   req.Browser.Channel,
		BrowserMode:      req.Browser.Mode,
		BrowserArgs:      req.Browser.Args,
		Tags:             req.Tags,
	})
	if err != nil {
		WriteError(c, err)
		return
	}

	c.JSON(http.StatusCreated, createSessionResponse{
		Session: toSessionResponse(view),
	})
}

func (s *Server) getSession(c *gin.Context) {
	if s.Sessions == nil {
		WriteError(c, errSessionServiceUnavailable)
		return
	}

	view, err := s.Sessions.Get(c.Request.Context(), tenantIDFromContext(c), c.Param("sessionId"))
	if err != nil {
		WriteError(c, err)
		return
	}

	c.JSON(http.StatusOK, toSessionResponse(view))
}

func (s *Server) deleteSession(c *gin.Context) {
	if s.Sessions == nil {
		WriteError(c, errSessionServiceUnavailable)
		return
	}

	view, err := s.Sessions.Delete(c.Request.Context(), tenantIDFromContext(c), c.Param("sessionId"))
	if err != nil {
		WriteError(c, err)
		return
	}

	c.JSON(http.StatusOK, sessionMutationResponse{
		Session: toSessionResponse(view),
	})
}

func (s *Server) replaceSessionTags(c *gin.Context) {
	if s.Sessions == nil {
		WriteError(c, errSessionServiceUnavailable)
		return
	}

	var req replaceTagsRequest
	if err := bindJSON(c, &req); err != nil {
		WriteError(c, err)
		return
	}

	view, err := s.Sessions.ReplaceTags(c.Request.Context(), tenantIDFromContext(c), c.Param("sessionId"), req.Tags)
	if err != nil {
		WriteError(c, err)
		return
	}

	c.JSON(http.StatusOK, sessionMutationResponse{
		Session: toSessionResponse(view),
	})
}

func (s *Server) suspendSession(c *gin.Context) {
	if s.Sessions == nil {
		WriteError(c, errSessionServiceUnavailable)
		return
	}

	view, err := s.Sessions.Suspend(c.Request.Context(), tenantIDFromContext(c), c.Param("sessionId"))
	if err != nil {
		WriteError(c, err)
		return
	}

	c.JSON(http.StatusOK, sessionMutationResponse{
		Session: toSessionResponse(view),
	})
}

func (s *Server) reopenSession(c *gin.Context) {
	if s.Sessions == nil {
		WriteError(c, errSessionServiceUnavailable)
		return
	}

	view, err := s.Sessions.Reopen(c.Request.Context(), tenantIDFromContext(c), c.Param("sessionId"))
	if err != nil {
		WriteError(c, err)
		return
	}

	c.JSON(http.StatusOK, sessionMutationResponse{
		Session: toSessionResponse(view),
	})
}